Move announcement components into a separate path
